requirements.txt
fire==0.4.0
python_telegram_bot==20.0
requests==2.22.0
